requirements.txt
pywebview==3.6.3
requests==2.28.1